refactor: DeleteComment Api 인터페이스와 구현체로 분리하여 구현
