+++ backend/footprint/src/main/java/foot/footprint/domain/comment/application/delete/DeleteCommentService.java
@@ -1,0 +1,8 @@
+package foot.footprint.domain.comment.application.delete;
+
+import foot.footprint.domain.article.exception.NotMatchMemberException;
+
+public interface DeleteCommentService {
+
+    public void delete(Long commentId, Long memberId);
+}
